xz1919: finished refactoring
