+++ src/main/kotlin/processing/filters/blur/LensBlur.kt
@@ -1,7 +1,8 @@
-package processing.filters
+package processing.filters.blur
 
 import javafx.scene.image.WritableImage
 import processing.ImageProcessing
+import processing.filters.Convolution
 import kotlin.math.pow
 
 class LensBlur(private val radius: Int) : ImageProcessing {
