+++ src/main/kotlin/processing/filters/blur/MotionBlur.kt
@@ -1,7 +1,8 @@
-package processing.filters
+package processing.filters.blur
 
 import javafx.scene.image.WritableImage
 import processing.ImageProcessing
+import processing.filters.Convolution
 
 class MotionBlur(private val radius: Int, private val angle: Double) : ImageProcessing {
     override fun process(image: WritableImage) {
