+++ src/main/kotlin/view/FilterPanel.kt
@@ -53,73 +53,24 @@ class FilterPanel : View() {
 
     override val root = vbox {
         splitpane {
-            scrollpane {
-                val tabPane = tabpane {
-                    tabClosingPolicy = TabPane.TabClosingPolicy.UNAVAILABLE
-                    side = Side.LEFT
-                    tab("Basic Actions") {
-                        content = BasicFilterTab(basicFilterButtonList)
-                    }
+            val tabPane = tabpane {
+                tabClosingPolicy = TabPane.TabClosingPolicy.UNAVAILABLE
+                side = Side.LEFT
+                tab("Basic Actions") {
+                    content = BasicFilterTab(basicFilterButtonList)
+                }
 
-                    tab("Color Adjust") {
-                        content = ColorAdjustTab(colorAdjustmentSliderList, engineController)
-                    }
+                tab("Color Adjust") {
+                    content = ColorAdjustTab(colorAdjustmentSliderList, engineController)
                 }
                 tab("frequency transfer") {
                     button("transfer").setOnAction {
                         engineController.frequencyTransfer()
                     }
                 }
-                tab("Blur & Sharpen") {
-                    vbox {
-                        label("Blur") {
-                            vboxConstraints {
-                                margin = Insets(20.0, 20.0, 10.0, 10.0)
-                            }
-                            style {
-                                fontWeight = FontWeight.BOLD
-                                fontSize = Dimension(20.0, Dimension.LinearUnits.px)
-                            }
-                        }
-                        vbox {
-                            vboxConstraints {
-                                margin = Insets(10.0)
-                            }
-                            val sliders = ArrayList<Slider>()
-                            hbox {
-                                padding = Insets(20.0, 20.0, 10.0, 10.0)
-                                val blurList = BlurType.values().toList()
-                                val combobox = combobox(values = blurList)
-                                combobox.value = blurList[0]
-                                val slider = slider {
-                                    min = 0.0
-                                    max = 10.0
-                                }
-                                sliders += slider
-                                slider.value = 0.0
-                                combobox.valueProperty()
-                                    .addListener(ChangeListener { _, _, _ ->
-                                        engineController.resetAdjustment()
-                                        slider.value = 0.0
-                                    })
-                                slider.valueChangingProperty()
-                                    .addListener(ChangeListener { _, _, _ ->
-                                        engineController.blur(slider.value.toInt(), combobox.value)
-                                    })
-
-                                addClass(CssStyle.filterSlider)
-                            }
-
-                    tab("Blur") {
-                        content = BlurFilterTab(engineController)
-                    }
+                tab("Blur") {
+                    content = BlurFilterTab(engineController)
                 }
-                tabPane.selectionModel.selectedIndexProperty()
-                    .addListener(ChangeListener<Number> { _, _, _ ->
-                        this.vvalue = 0.0
-                    })
-                vbarPolicy = ScrollPane.ScrollBarPolicy.NEVER
-                hbarPolicy = ScrollPane.ScrollBarPolicy.NEVER
             }
 
             vbox {
